# [Kernel] Improve machete memory bound perf